chore: refactored to separate HTTP exceptions (#2456)
