+++ backend/corpora/lambdas/api/v1/collection.py
@@ -13,7 +13,7 @@ from .common import get_collection
 from ....common.corpora_orm import DbCollection, CollectionVisibility
 from ....common.entities import Collection
 from .authorization import is_user_owner_or_allowed, owner_or_allowed
-from ....common.utils.exceptions import (
+from ....common.utils.http_exceptions import (
     InvalidParametersHTTPException,
     ConflictException,
 )
