+++ backend/corpora/lambdas/api/v1/collection_uuid/publish.py
@@ -2,10 +2,10 @@ from flask import make_response, g
 
 from .....common.corpora_orm import CollectionVisibility
 from .....common.entities import Collection
-from .....common.utils.exceptions import ConflictException
+from .....common.utils.http_exceptions import ConflictException
 
 from .....api_server.db import dbconnect
-from .....common.utils.exceptions import ForbiddenHTTPException
+from .....common.utils.http_exceptions import ForbiddenHTTPException
 from ..authorization import owner_or_allowed
 
 from backend.corpora.common.utils import cloudfront
