+++ tests/unit/backend/corpora/common/utils/test_color_conversion_utils.py
@@ -6,7 +6,7 @@ from backend.corpora.common.utils.color_conversion_utils import (
     convert_color_to_hex_format,
     convert_anndata_category_colors_to_cxg_category_colors,
 )
-from backend.corpora.common.utils.exceptions import ColorFormatException
+from backend.corpora.common.utils.http_exceptions import ColorFormatException
 from tests.unit.backend.corpora.fixtures.environment_setup import fixture_file_path
